refactor: separar DoctorService em interface e classe de implementação
